Add files via upload
The collatz operations function without using a cache and with using an LRU cache. The LFU cache is currently unusable. Organizing the LFU cache based on frequency has not been implemented yet. The collatzTimeTest.sh file can be used to output the collatz values in one csv file and the time, hit ratio, and cache details in a separate csv file.